Generate the results table

Contains the logic to generate the results table

Usage

orbi_calculate_results(dataset, ratio.method)

Arguments

dataset

A processed tibble produced from IsoX output

ratio.method

Method for computing the ratio; passed to orbi_calculate_ratio()

Value

Returns table containing filename, compound, Basepeak, Isotopocule, Ratio, Ratio.SEM, relSE.permil, Shot.Noise.permil, No.of.Scans, Mins.to.1mio

Details

Please note well: The formula used to calculate ion ratios matters! Do not simply use arithmetic mean. The best option may depend on the type of data you are processing (e.g., MS1 versus M+1 fragmentation).

Description of options for ratio.method:

mean: arithmetic mean of ratios from individual scans.

sum: sum of all ions of the numerator across all scans divided by the sum of all ions observed for the denominator across all scans.

geometric.mean: geometric mean of ratios from individual scans.

slope: The ratio is calculated using the slope obtained from a linear regression model that is weighted by the numerator x, using stats::lm(x ~ y + 0, weights = x).

weighted.sum: A derivative of the sum option. The weighing function ensures that each scan contributes equal weight to the ratio calculation, i.e. scans with more ions in the Orbitrap do not contribute disproportionately to the total sum of x and y that is used to calculate x/y.

Examples

fpath <- system.file("extdata", "testfile_Flow_Exploration_small.isox", package = "isoorbi")
df <- orbi_read_isox(filepath = fpath) %>%
      orbi_simplify_isox() %>% orbi_basepeak(basepeak = "M0")  %>%
      orbi_calculate_results(ratio.method = "sum")