Flag outlier scans

The function orbi_flag_outliers() flags outliers. Grouping is by columns filename and compound.

Usage

orbi_flag_outliers(dataset, intensity_window)

Arguments

dataset

Simplified IsoX dataset to have outliers flagged

intensity_window

flags extremely high and low intense scans based on TIC x injection time (i.e., ion intensity). Provide a vector with 2 numbers c(x,y) filtering the lowest x percent intensities and highest y percent intensities.

Value

Filtered tibble

Details

Function is intended to flag scans that are outliers. TIC multiplied by injection time serves as an estimate for the number of ions in the Orbitrap.

The input dataset is expected to have at least these 8 columns: filename, scan.no, time.min, compound, isotopocule, ions.incremental, tic, it.ms.

Examples

fpath <- system.file("extdata", "testfile_flow.isox", package = "isoorbi")
df <-
  orbi_read_isox(file = fpath) |>
  orbi_simplify_isox() |>
  orbi_flag_outliers(intensity_window = c(1,99))
#> orbi_read_isox() is loading .isox data from 1 file(s)...
#> - loaded 6449 peaks for 1 compounds (HSO4-) with 5 isotopocules (M0, 33S,
#>    17O, 34S, 18O) from testfile_flow.isox in 0.02 seconds.
#> orbi_simplify_isox() will keep only columns 'filename', 'compound',
#>    'scan.no', 'time.min', 'isotopocule', 'ions.incremental', 'tic',
#>    'it.ms'...
#>    ...complete in 0.01 seconds.
#> orbi_filter_scan_intensity() is flagging the 1 % of scans with the lowest and
#>    above 99 % of scans with the highest intensities in 3 data group(s) (based
#>    on 'filename')...
#>    ...flagged 30/1290 scans (2.3%) across all data groups in 0.05 seconds.