Identify isotopocules

Map the mass spectral peaks to specific isotopocules based on their mass.

Usage

orbi_identify_isotopocules(aggregated_data, isotopocules)

Arguments

aggregated_data

either data aggregated from orbi_aggregate_raw() or a straight-up tibble data frame of the peaks (e.g. retrieved via orbi_get_data(peaks = everything())).

isotopocules

list of isotopocules to map, can be a data frame/tibble or name of a file to read from (.csv/.tsv/.xlsx are all supported). Required columns are isotopocule/isotopolog, mz/mass, tolerance/tolerance [mmu]/tolerance [mDa], and charge/z (these alternative names for the columns, including uppercase versions, are recognized automatically). Optional column: #compound/compound as well as any other columns that don't match these others. Character columns in the isotopocules table (including isotopocule and compound) are turned into factors with levels that preserve the order of isotopocules. That means that to change the order of isotopocules in downstream plotting functions, make sure to list them in the order you'd like them presented in.

Value

same object as provided in aggregated_data with added columns compound (if provided), itc_uidx (introduced unique isotopocule index), isotopocule, mzExact and charge as well as any other additional information columns provided in isotopocules. Note that the information about columns that were NOT aggregated in previous steps is purposefully not preserved at this step.