Package options

These options are best set via orbi_options() and queried via orbi_get_option(). However, the base functions options() and getOption() work as well but require an isoorbi. prefix (the package name and a dot) for the option name. Setting an option to a value of NULL means that the default is used. orbi_get_options() is available as an additional convenience function to retrieve a subset of options with a regular expression pattern.

Usage

orbi_options(...)

orbi_get_options(pattern = NULL)

orbi_get_option(x)

Arguments

...

set package options, syntax identical to options()

pattern

to retrieve multiple options (as a list) with a shared pattern

x

name of the specific option to retrieve

Functions

orbi_options(): set/get option values

orbi_get_options(): get a subset of option values that fit a pattern

orbi_get_option(): retrieve the current value of one option (option must be defined for the package)

Options for the isoorbi package

di_ref_name: the text label for dual inlet reference blocks

di_sample_name: the text label for dual inlet sample blocks

data_type_data: the text used to flag raw data as actually being data

data_type_startup: the text used to flag raw data as being part of the startup

data_type_changeover: the text used to flag raw data as being part of a changeover

data_type_unused: the text used to flag raw data as being unused

raw_aggregator: configuration for pulling data out of raw files

debug: turn on debug mode

auto_use_ansi: whether to automatically enable correct rendering of stylized (ansi) output in HTML reports from notebooks that call library(isoorbi). Can be turned off by calling isoorbi::orbi_options(auto_use_ansi = FALSE) before call library(isoorbi).

Examples

# All default options
orbi_get_options()
#> $di_ref_name
#> [1] "ref"
#> 
#> $di_sample_name
#> [1] "sam"
#> 
#> $data_type_data
#> [1] "data"
#> 
#> $data_type_startup
#> [1] "startup"
#> 
#> $data_type_changeover
#> [1] "changeover"
#> 
#> $data_type_unused
#> [1] "unused"
#> 
#> $raw_aggregator
#> # A tibble: 22 × 8
#>    dataset   column              source     default regexp cast     func  args  
#>    <chr>     <chr>               <list>     <lgl>   <lgl>  <chr>    <chr> <list>
#>  1 file_info "uid"               <list [1]> NA      FALSE  as.fact… NA    <NULL>
#>  2 file_info "\\1"               <list [1]> NA      TRUE   as.char… NA    <NULL>
#>  3 scans     "scan"              <list [1]> NA      FALSE  as.inte… NA    <NULL>
#>  4 scans     "time.min"          <list [1]> NA      FALSE  as.nume… NA    <NULL>
#>  5 scans     "tic"               <list [1]> NA      FALSE  as.nume… NA    <NULL>
#>  6 scans     "it.ms"             <list [1]> NA      FALSE  as.nume… NA    <NULL>
#>  7 scans     "resolution"        <list [1]> NA      FALSE  as.nume… NA    <NULL>
#>  8 scans     "basePeakIntensity" <list [1]> NA      FALSE  as.nume… sapp… <list>
#>  9 scans     "rawOvFtT"          <list [1]> NA      FALSE  as.nume… NA    <NULL>
#> 10 scans     "intensCompFactor"  <list [1]> NA      FALSE  as.nume… NA    <NULL>
#> # ℹ 12 more rows
#> 
#> $debug
#> [1] FALSE
#> 
#> $auto_use_ansi
#> [1] TRUE
#> 

# Options that contain 'data' in the name
orbi_get_options("data")
#> $data_type_data
#> [1] "data"
#> 
#> $data_type_startup
#> [1] "startup"
#> 
#> $data_type_changeover
#> [1] "changeover"
#> 
#> $data_type_unused
#> [1] "unused"
#> 

# Specific option
orbi_get_option("data_type_unused")
#> [1] "unused"

# Change an option
orbi_options(data_type_unused = "flagged")
orbi_get_option("data_type_unused")
#> [1] "flagged"

# Change back to default
orbi_options(data_type_unused = NULL)
orbi_get_option("data_type_unused")
#> [1] "unused"