Retrieve parsing problems

This function retrieves parsing problems encountered during the reading and processing of files.

This function prints out parsing problems encountered during the reading and processing of files.

Usage

orbi_get_problems(obj, strip_ansi = TRUE)

orbi_show_problems(obj)

Arguments

obj

data object that holds problems information

strip_ansi

whether to remove ansi characters from the message, yes by default

Value

tibble data frame with a list of problems encountered during

Examples

orbi_find_raw(system.file("extdata", package = "isoorbi"))[1] |>
  orbi_read_raw() |>
  orbi_get_problems()
#> ! Could not find the executable for the RAW file reader:
#>   'rawrr.exe' not found. Run 'rawrr::installRawrrExe()'. For more information,
#>   type '?rawrr.exe'.
#> → Trying to install rawrr.exe (this requires an internet connection and may
#>   take a moment)...
#> Overwrite sourceUrl to https://fgcz-ms.uzh.ch/~cpanse/rawrr/dotnet//linux-x64/rawrr
#> MD5 c51e3eb46ba5633b9763e94e9d09769c /home/runner/.cache/R/rawrr/rawrrassembly/linux-x64/rawrr
#> ✔ [772ms] orbi_read_raw() read nitrate_test_10scans.raw (124.68 kB)
#> # A tibble: 0 × 5
#> # ℹ 5 variables: uid <fct>, type <chr>, call <chr>, message <chr>,
#> #   condition <list>

orbi_find_raw(system.file("extdata", package = "isoorbi"))[1] |>
  orbi_read_raw() |>
  orbi_show_problems()
#> ✔ [32ms] orbi_read_raw() read nitrate_test_10scans.raw from cache