Flag minor satellite peaks

Flag minor signals (e.g., satellite peaks) that were reported by IsoX (filter them out with orbi_filter_flagged_data()).

Usage

orbi_flag_satellite_peaks(dataset)

Arguments

dataset

A data frame or tibble produced from IsoX data by orbi_simplify_isox()

Value

A data frame with new column is_satellite_peak that flags satellite peaks.

Details

The orbi_filter_satellite_peaks() function removes minor signals for an isotopocule that have been reported by IsoX. These are often small satellite peaks generated by the Fourier transform.

If there are signal of high intensity or very many signals, this can indicate that the m/z and tolerance setting used for processing .raw files with IsoX were incorrect.

Examples

fpath <- system.file("extdata", "testfile_flow.isox", package = "isoorbi")
df <-
  orbi_read_isox(file = fpath) |>
  orbi_simplify_isox() |>
  orbi_flag_satellite_peaks()
#> ✔ [20ms] orbi_read_isox() loaded 6449 peaks for 1 compound (HSO4-) with 5
#> isotopocules (M0, 33S, 17O, 34S, and 18O) from testfile_flow.isox
#> ✔ [5ms] orbi_simplify_isox() kept columns filepath, filename, scan.no,
#> time.min, compound, isotopocule, ions.incremental, tic, and it.ms
#> ✔ [173ms] orbi_flag_satellite_peaks() confirmed there are no satellite peaks