Generate the results table

Contains the logic to generate the results table

Usage

orbi_summarize_results(dataset, ratio_method)

Arguments

dataset

A processed tibble produced from IsoX output

ratio_method

Method for computing the ratio; passed to orbi_calculate_ratio()

Value

Returns a results table containing filename, compound, basepeak, Isotopocule, ratio, ratio_sem, ratio_relative_sem_permil, shot_noise_permil, No.of.Scans, minutes_to_1e6_ions

Details

Description of the output columns:

basepeak: Isotopocule used as denominator in ratio calculation.

isotopocule: Isotopocule used as numerator in ratio calculation.

ratio_sem: Standard error of the mean for the ratio

number_of_scans: Number of scans used for the final ratio calculation

minutes_to_1e6_ions: Time in minutes it would take to observe 1 million ions of the isotopocule used as numerator of the ratio calculation.

shot_noise_permil: Estimate of the shot noise (more correctly thermal noise) of the reported ratio in permil.

ratio_relative_sem_permil: Relative standard error of the reported ratio in permil

Examples

fpath <- system.file("extdata", "testfile_Flow_Exploration_small.isox", package = "isoorbi")
df <- orbi_read_isox(filepath = fpath) %>%
      orbi_simplify_isox() %>% orbi_define_basepeak(base_peak = "M0")  %>%
      orbi_summarize_results(ratio_method = "sum")
#> orbi_read_isox() is loading .isox data from file path: 
#> /home/runner/work/_temp/Library/isoorbi/extdata/testfile_Flow_Exploration_small.isox
#> orbi_simplify_isox() is applied to dataset .. Will keep only most important columns...
#> orbi_define_basepeak() is setting the M0 isotopocule as the ratio denominator...
#> orbi_summarize_results() is grouping dataset by columns: filename, compound, basepeak, isotopocule
#> orbi_summarize_results() is adding a grouping: block
#> orbi_summarize_results() is adding a grouping: segment
#> orbi_summarize_results() is adding a grouping: injection
#> orbi_calculate_ratio() is calculating ratios, using the ratio_method: sum
#> orbi_summarize_results() is summarizing ratios for the results table...